trimmer/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
asgiref=3.2.3=py_0
beautifulsoup4=4.8.2=pypi_0
ca-certificates=2019.11.28=hecc5488_0
certifi=2019.11.28=py38_0
django=3.0.2=py38_0
django-bootstrap4=1.1.1=pypi_0
django-filter=2.2.0=pypi_0
django-widget-tweaks=1.4.5=pypi_0
libcxx=4.0.1=hcfea43d_1
libcxxabi=4.0.1=hcfea43d_1
libedit=3.1.20181209=hb402a30_0
libffi=3.2.1=h475c297_4
mysqlclient=1.4.6=pypi_0
ncurses=6.1=h0a44026_1
openssl=1.1.1d=h0b31af3_0
pip=19.3.1=py38_0
python=3.8.1=h359304d_1
pytz=2019.3=py_0
readline=7.0=h1de35cc_5
setuptools=44.0.0=py38_0
soupsieve=1.9.5=pypi_0
sqlite=3.30.1=ha441bb4_0
sqlparse=0.3.0=py_0
tk=8.6.8=ha441bb4_0
wheel=0.33.6=py38_0
xz=5.2.4=h1de35cc_4
zlib=1.2.11=h1de35cc_3
